armorbotservice/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0
langchain==0.0.215
pinecone-client==2.2.2
sumy==0.11.0
openai==0.27.8
beautifulsoup4==4.12.2
tiktoken==0.4.0
html2text==2020.1.16
PyPDF2==3.0.1
grequests==0.7.0
validators==0.20.0
urllib3==2.0.3
